Updated formatting for logger debug so line # will be 5 space right-aligned. Change to use alternative entrypoint. Refactored the main error handling by itself as a pre-check
