+++ .gitignore
@@ -89,3 +89,5 @@ test-output.xml
 
 # vim swap
 *.swp
+
+todo.txt
