+++ SSLChecker/SSLChecker/function.json
@@ -1,5 +1,6 @@
 {
-  "scriptFile": "__init__.py",
+  "scriptFile": "main.py",
+  "entryPoint": "main",
   "bindings": [
     {
       "authLevel": "anonymous",
