+++ SSLChecker/sharedcode/errors.py
@@ -1,5 +1,5 @@
 """
-Raised when config.ini is not valid
+Custom Exception for the app
 """
 class InvalidConfig(Exception):
     pass
@@ -8,4 +8,10 @@ class InvalidFQDN(Exception):
     pass
 
 class UnknownError(Exception):
+    pass
+
+class InvalidRequest(Exception):
+    pass
+
+class DNSError(Exception):
     pass
